fix: bluesky time-bound query issues
The `Bluesky` bundle's `Search Post` tool fails to search for posts when
the `since` or `until` parameters are set due an error returned by the
public bsky search APIs. This isn't well documented, but after some
testing it was determined that this only occurs when making
unauthenticated requests.

To fix this issue, this change updates the Bluesky tools to make
authenticated requests with a user handle and [App Password](https://bsky.app/settings/app-passwords)
provided by the user via a basic auth credential tool.

This change also exposes a `tags` parameter on the `Search Post` tool to
make it easier for agents to search by tag.

Signed-off-by: Nick Hale <4175918+njhale@users.noreply.github.com>